+++ libraries/common/src/main/java/androidx/media3/common/Player.java
@@ -3207,6 +3207,16 @@ public interface Player {
   AudioAttributes getAudioAttributes();
 
   /**
+   * Returns the audio session identifier, or {@link C#AUDIO_SESSION_ID_UNSET} if not set.
+   *
+   * @see Listener#onAudioSessionIdChanged(int)
+   */
+  @UnstableApi
+  default int getAudioSessionId() {
+    return C.AUDIO_SESSION_ID_UNSET;
+  }
+
+  /**
    * Sets the audio volume, valid values are between 0 (silence) and 1 (unity gain, signal
    * unchanged), inclusive.
    *