+++ libraries/session/src/main/java/androidx/media3/session/MediaSessionImpl.java
@@ -1866,6 +1866,24 @@ import org.checkerframework.checker.initialization.qual.Initialized;
     }
 
     @Override
+    public void onAudioSessionIdChanged(int audioSessionId) {
+      @Nullable MediaSessionImpl session = getSession();
+      if (session == null) {
+        return;
+      }
+      session.verifyApplicationThread();
+      @Nullable PlayerWrapper player = this.player.get();
+      if (player == null) {
+        return;
+      }
+      session.playerInfo = session.playerInfo.copyWithAudioSessionId(audioSessionId);
+      session.onPlayerInfoChangedHandler.sendPlayerInfoChangedMessage(
+          /* excludeTimeline= */ true, /* excludeTracks= */ true);
+      session.dispatchRemoteControllerTaskToLegacyStub(
+          (callback, seq) -> callback.onAudioSessionIdChanged(seq, audioSessionId));
+    }
+
+    @Override
     public void onAudioAttributesChanged(AudioAttributes attributes) {
       @Nullable MediaSessionImpl session = getSession();
       if (session == null) {