+++ libraries/session/src/main/java/androidx/media3/session/PlayerWrapper.java
@@ -938,6 +938,7 @@ import java.util.List;
         getVolumeWithCommandCheck(),
         /* unmuteVolume= */ 1f,
         getAudioAttributesWithCommandCheck(),
+        C.AUDIO_SESSION_ID_UNSET,
         getCurrentCuesWithCommandCheck(),
         getDeviceInfo(),
         getDeviceVolumeWithCommandCheck(),