+++ RELEASENOTES.md
@@ -176,6 +176,9 @@
         commands are sent to the session only if the custom action is advertised
         as a custom action in `PlaybackStateCompat` of the legacy session. All
         other custom actions are sent to the service.
+    *   Implemented `onAudioSessionIdChanged` to notify media controllers when
+        an audio session ID is set by the session
+        ([#244](https://github.com/androidx/media/issues/244)).
 *   UI:
     *   Add `ProgressStateWithTickInterval` class and the corresponding
         `rememberProgressStateWithTickInterval` Composable to